+++ internal/mapper/v04/silences.go
@@ -14,7 +14,7 @@ import (
 	"github.com/blang/semver"
 	"github.com/cloudflare/unsee/internal/mapper"
 	"github.com/cloudflare/unsee/internal/models"
-	"github.com/cloudflare/unsee/internal/transport"
+	"github.com/cloudflare/unsee/internal/uri"
 )
 
 // Alertmanager 0.4 silence format
@@ -49,7 +49,7 @@ type SilenceMapper struct {
 
 // AbsoluteURL for silences API endpoint this mapper supports
 func (m SilenceMapper) AbsoluteURL(baseURI string) (string, error) {
-	return transport.JoinURL(baseURI, "api/v1/silences")
+	return uri.JoinURL(baseURI, "api/v1/silences")
 }
 
 // IsSupported returns true if given version string is supported
