+++ internal/uri/file.go
@@ -1,4 +1,4 @@
-package transport
+package uri
 
 import (
 	"io"
